auth & guest middleware

+++ routes/web.php
@@ -19,30 +19,30 @@ use App\Http\Controllers\ListingController;
 
 Route::get('/', [ListingController::class, 'index']);
 
-Route::get('/listings/create', [ListingController::class,'create']);
+Route::get('/listings/create', [ListingController::class,'create'])->middleware('auth');
 
-Route::post('/listings', [ListingController::class,'store']);
+Route::post('/listings', [ListingController::class,'store'])->middleware('auth');
 
-Route::get('/listings/{listing}/edit', [ListingController::class,'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class,'edit'])->middleware('auth');
 
 // Update post
-Route::put('/listings/{listing}', [ListingController::class,'update']);
+Route::put('/listings/{listing}', [ListingController::class,'update'])->middleware('auth');
 
 //Delete post
-Route::delete('/listings/{listing}', [ListingController::class,'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class,'destroy'])->middleware('auth');
 
 Route::get('/listings/{listing}', [ListingController::class,'show']);
 
 // register user
-Route::get('/register', [UserController::class,'create']);
+Route::get('/register', [UserController::class,'create'])->middleware('guest');
 
 Route::post('/users', [UserController::class,'store']);
 
 //logout user
-Route::post('/logout', [UserController::class,'logout']);
+Route::post('/logout', [UserController::class,'logout'])->middleware('auth');
 
 //show login form
-Route::get('/login', [UserController::class,'login']);
+Route::get('/login', [UserController::class,'login'])->name('login')->middleware('auth');
 
 //Login user
 Route::post('/users/authenticate', [UserController::class,'authenticate']);
